Example of making network calls to the backend/server
Also moved around a few components/functions
